+++ CHANGES.md
@@ -4,6 +4,11 @@
 
 - Do not normalize sources in `mix` by default.
 
+## Fixed:
+
+- Raise a proper error on parse error raised during `%include`
+  parsing.
+
 ---
 
 # 2.4.0 (2025-09-01)